Moved role policy management classes to common
